feat: change components structure
