+++ src/components/WorkLog/WorkLogDifficulties/WorkLogDifficulties.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
-import data from '../../data/difficulties.json';
+import data from './difficulties.json';
 
 const WorkLogDifficulties = ({ language }) => {
   return (
     <div>
       <h3>{data[language].title}</h3>
-      <p>{data[language].text}</p>
+      <p>{data[language].text1}</p>
     </div>
   );
 };
