+++ src/components/WorkLog/WorkLogTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from './WorkLogTable';
