--- src/data/difficulties.json
@@ -1,14 +1,0 @@
-{
-  "ru": {
-    "title": "Трудности в нашей работе",
-    "text": "Основной трудностью в нашей работе являлась, вероятно, организация взаимодействия между собой членов команды. И, можно сказать, эту трудность мы сообща преодолели. Каждый внес свой вклад в общее дело. И сегодня вы можете увидеть этот портал, посвященный писателям Беларуси."
-  },
-  "en": {
-    "title": "Work Difficulties",
-    "text": "The main difficulty in our work was probably the organization of interaction between team members. And, we can say, we have overcome this difficulty together. Everyone has contributed to the common cause. And today you can see this portal dedicated to the writers of Belarus."
-  },
-  "be": {
-    "title": "Цяжкасці ў нашай працы",
-    "text": "Асноўнай цяжкасцю у нашай працы з'яўлялася, верагодна, арганізацыя ўзаемадзеяння паміж сабой членаў каманды. І, можна сказаць, гэтую цяжкасць мы разам пераадолелі. Кожны ўнёс свой уклад у агульную справу. І сёння вы можаце ўбачыць гэты партал, прысвечаны пісьменнікам Беларусі."
-  }
-}
